Code review returns

+++ server/src/metadata/workspace-migration/migrations/1697618021-addOpportunityTable.ts
@@ -14,7 +14,7 @@ export const addOpportunityTable: WorkspaceMigrationTableAction[] = [
     columns: [
       {
         columnName: 'amountAmountMicros',
-        columnType: 'bigint',
+        columnType: 'numeric',
         action: WorkspaceMigrationColumnActionType.CREATE,
       },
       {
